+++ app/components/InspectorContent.vue
@@ -1,7 +1,7 @@
 <template>
   <div
     data-testid="inspector-content"
-    class="flex min-h-0 flex-1 flex-col gap-2 overflow-y-auto p-5"
+    class="inspector-content-shell flex min-h-0 flex-1 flex-col overflow-y-auto"
   >
     <InspectorPropertiesView />
   </div>