+++ app/components/NoteEditor.vue
@@ -1,6 +1,6 @@
 <template>
   <div
     data-testid="note-editor"
-    class="flex min-h-0 flex-1 flex-col bg-card px-[100px] pt-[50px]"
+    class="note-editor-shell flex min-h-0 flex-1 flex-col"
   />
 </template>