+++ app/components/NoteTemplate.vue
@@ -1,7 +1,7 @@
 <template>
   <div
     data-testid="note-template"
-    class="flex min-h-0 flex-1 flex-col overflow-hidden rounded-t-[10px] bg-[#3e56de12] px-[100px] pt-[50px]"
+    class="note-template-shell flex min-h-0 flex-1 flex-col overflow-hidden"
   >
     <NoteEditor />
   </div>